requirements.txt
mss
Pillow
PyAudio
openai
google-generativeai
python-dotenv
keyboard
